Kernel/PCI: Move the PCI components as a subfolder to the Bus directory
