+++ Kernel/Bus/PCI/Access.cpp
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <Kernel/Bus/PCI/Access.h>
+#include <Kernel/Bus/PCI/IOAccess.h>
 #include <Kernel/Debug.h>
 #include <Kernel/IO.h>
-#include <Kernel/PCI/Access.h>
-#include <Kernel/PCI/IOAccess.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
