+++ Kernel/Bus/PCI/Access.h
@@ -9,8 +9,8 @@
 #include <AK/Bitmap.h>
 #include <AK/String.h>
 #include <AK/Vector.h>
+#include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/FileSystem/SysFS.h>
-#include <Kernel/PCI/Definitions.h>
 
 namespace Kernel::PCI {
