+++ Kernel/Bus/PCI/Device.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/PCI/Device.h>
+#include <Kernel/Bus/PCI/Device.h>
 
 namespace Kernel {
 namespace PCI {
