+++ Kernel/Bus/PCI/WindowedMMIOAccess.h
@@ -11,8 +11,8 @@
 #include <AK/OwnPtr.h>
 #include <AK/Types.h>
 #include <Kernel/ACPI/Definitions.h>
-#include <Kernel/PCI/Access.h>
-#include <Kernel/PCI/MMIOAccess.h>
+#include <Kernel/Bus/PCI/Access.h>
+#include <Kernel/Bus/PCI/MMIOAccess.h>
 #include <Kernel/VM/AnonymousVMObject.h>
 #include <Kernel/VM/PhysicalRegion.h>
 #include <Kernel/VM/Region.h>
