+++ Kernel/Graphics/GraphicsDevice.h
@@ -8,8 +8,8 @@
 
 #include <AK/String.h>
 #include <AK/Types.h>
+#include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/Devices/BlockDevice.h>
-#include <Kernel/PCI/Definitions.h>
 #include <Kernel/PhysicalAddress.h>
 
 namespace Kernel {
