+++ Kernel/Graphics/GraphicsManagement.h
@@ -10,11 +10,11 @@
 #include <AK/NonnullRefPtr.h>
 #include <AK/NonnullRefPtrVector.h>
 #include <AK/Types.h>
+#include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/Graphics/Console/Console.h>
 #include <Kernel/Graphics/GraphicsDevice.h>
 #include <Kernel/Graphics/VGACompatibleAdapter.h>
 #include <Kernel/Graphics/VirtIOGPU/VirtIOGraphicsAdapter.h>
-#include <Kernel/PCI/Definitions.h>
 #include <Kernel/VM/Region.h>
 
 namespace Kernel {
