+++ Kernel/Graphics/IntelNativeGraphicsAdapter.h
@@ -8,10 +8,10 @@
 
 #include <AK/String.h>
 #include <AK/Types.h>
+#include <Kernel/Bus/PCI/DeviceController.h>
 #include <Kernel/Graphics/Definitions.h>
 #include <Kernel/Graphics/FramebufferDevice.h>
 #include <Kernel/Graphics/VGACompatibleAdapter.h>
-#include <Kernel/PCI/DeviceController.h>
 #include <Kernel/PhysicalAddress.h>
 
 namespace Kernel {
