+++ Kernel/Graphics/VGACompatibleAdapter.h
@@ -8,10 +8,10 @@
 
 #include <AK/String.h>
 #include <AK/Types.h>
+#include <Kernel/Bus/PCI/DeviceController.h>
 #include <Kernel/Graphics/Console/Console.h>
 #include <Kernel/Graphics/FramebufferDevice.h>
 #include <Kernel/Graphics/GraphicsDevice.h>
-#include <Kernel/PCI/DeviceController.h>
 #include <Kernel/PhysicalAddress.h>
 
 namespace Kernel {
