+++ Kernel/Net/E1000ENetworkAdapter.cpp
@@ -5,9 +5,9 @@
  */
 
 #include <AK/MACAddress.h>
+#include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Net/E1000ENetworkAdapter.h>
-#include <Kernel/PCI/IDs.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
