+++ Kernel/Net/E1000NetworkAdapter.cpp
@@ -5,9 +5,9 @@
  */
 
 #include <AK/MACAddress.h>
+#include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Net/E1000NetworkAdapter.h>
-#include <Kernel/PCI/IDs.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
