+++ Kernel/Net/NetworkAdapter.h
@@ -13,12 +13,12 @@
 #include <AK/Types.h>
 #include <AK/WeakPtr.h>
 #include <AK/Weakable.h>
+#include <Kernel/Bus/PCI/Definitions.h>
 #include <Kernel/KBuffer.h>
 #include <Kernel/Net/ARP.h>
 #include <Kernel/Net/EthernetFrameHeader.h>
 #include <Kernel/Net/ICMP.h>
 #include <Kernel/Net/IPv4.h>
-#include <Kernel/PCI/Definitions.h>
 #include <Kernel/UserOrKernelBuffer.h>
 
 namespace Kernel {
