+++ Kernel/Net/RTL8168NetworkAdapter.cpp
@@ -5,10 +5,10 @@
  */
 
 #include <AK/MACAddress.h>
+#include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/IO.h>
 #include <Kernel/Net/RTL8168NetworkAdapter.h>
-#include <Kernel/PCI/IDs.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
