+++ Kernel/Storage/StorageManagement.cpp
@@ -5,10 +5,10 @@
  */
 
 #include <AK/UUID.h>
+#include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Devices/BlockDevice.h>
 #include <Kernel/FileSystem/Ext2FileSystem.h>
-#include <Kernel/PCI/Access.h>
 #include <Kernel/Panic.h>
 #include <Kernel/Storage/AHCIController.h>
 #include <Kernel/Storage/IDEController.h>
